+++ ViroRenderer/giflib/dgif_lib.c
@@ -1135,7 +1135,7 @@ DGifSlurp(GifFileType *GifFile)
 		      }
 	      }
 	      else {
-		  if (DGifGetLine(GifFile,sp->RasterBits,ImageSize)==GIF_ERROR)
+		  if (DGifGetLine(GifFile,sp->RasterBits, (int) ImageSize)==GIF_ERROR)
 		      return (GIF_ERROR);
 	      }
